+++ tests/testBinIO.cpp
@@ -61,7 +61,7 @@ TEST(BinIO_testWriteFile) {
 	Buffer buf;
 	BufferStatus status = bufInitialise(&buf, 8, 0, NULL);
 	CHECK_EQUAL(BUF_SUCCESS, status);
-	status = bufAppendBlock(&buf, (const uint8 *)DATA, (uint32)strlen(DATA), NULL);
+	status = bufAppendBlock(&buf, (const uint8 *)DATA, strlen(DATA), NULL);
 	CHECK_EQUAL(BUF_SUCCESS, status);
 
 	status = bufWriteBinaryFile(&buf, FILENAME, 0, buf.length, NULL);
